requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.11.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
Django==5.0
django-debug-toolbar==4.2.0
django-filter==23.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
flake8==6.1.0
inflection==0.5.1
isort==5.13.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.13.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
